+++ crates/fluvio-storage/src/config.rs
@@ -2,7 +2,7 @@ use std::default::Default;
 use std::path::PathBuf;
 use std::fmt;
 use std::sync::Arc;
-use std::sync::atomic::AtomicU32;
+use std::sync::atomic::{AtomicU32, AtomicU64};
 
 use derive_builder::Builder;
 use fluvio_controlplane_metadata::partition::Replica;
@@ -11,11 +11,11 @@ use serde::Deserialize;
 
 use fluvio_types::defaults::{
     SPU_LOG_INDEX_MAX_BYTES, SPU_LOG_BASE_DIR, STORAGE_FLUSH_WRITE_COUNT, STORAGE_FLUSH_IDLE_MSEC,
-    STORAGE_MAX_BATCH_SIZE, STORAGE_RETENTION_SECONDS,
+    STORAGE_MAX_BATCH_SIZE, STORAGE_RETENTION_SECONDS, SPU_PARTITION_MAX_BYTES,
 };
 use fluvio_types::defaults::SPU_LOG_INDEX_MAX_INTERVAL_BYTES;
 use fluvio_types::defaults::SPU_LOG_SEGMENT_MAX_BYTES;
-use dataplane::Size;
+use dataplane::{Size, Size64};
 
 use crate::{ReplicaStorageConfig};
 
@@ -50,6 +50,9 @@ pub struct ReplicaConfig {
     #[builder(default = "default_retention_seconds()")]
     #[serde(default = "default_retention_seconds")]
     pub retention_seconds: Size,
+    #[builder(default = "default_max_partition_size()")]
+    #[serde(default = "default_max_partition_size")]
+    pub max_partition_size: Size64,
 }
 
 impl fmt::Display for ReplicaConfig {
@@ -74,6 +77,13 @@ impl ReplicaStorageConfig for ReplicaConfig {
             .and_then(|storage| storage.segment_size)
         {
             self.segment_max_bytes = segment_size;
+        }
+        if let Some(max_partition_size) = replica
+            .storage
+            .as_ref()
+            .and_then(|storage| storage.max_partition_size)
+        {
+            self.max_partition_size = max_partition_size;
         }
     }
 }
@@ -114,6 +124,10 @@ const fn default_retention_seconds() -> Size {
     STORAGE_RETENTION_SECONDS
 }
 
+const fn default_max_partition_size() -> Size64 {
+    SPU_PARTITION_MAX_BYTES
+}
+
 impl ReplicaConfig {
     // Used to get a [`ConfigOptionBuilder`].
     pub fn builder() -> ReplicaConfigBuilder {
@@ -136,6 +150,7 @@ impl Default for ReplicaConfig {
             flush_idle_msec: default_flush_idle_msec(),
             max_batch_size: default_max_batch_size(),
             retention_seconds: default_retention_seconds(),
+            max_partition_size: default_max_partition_size(),
             update_hw: true,
         }
     }
@@ -153,9 +168,9 @@ impl ReplicaConfigBuilder {
 }
 
 #[derive(Debug)]
-pub struct SharedConfigValue(AtomicU32);
+pub struct SharedConfigValue<V: Sync + Send>(V);
 
-impl SharedConfigValue {
+impl SharedConfigValue<AtomicU32> {
     pub fn new(value: u32) -> Self {
         SharedConfigValue(AtomicU32::new(value))
     }
@@ -176,32 +191,58 @@ impl SharedConfigValue {
     }
 }
 
+impl SharedConfigValue<AtomicU64> {
+    pub fn new(value: u64) -> Self {
+        SharedConfigValue(AtomicU64::new(value))
+    }
+
+    #[inline(always)]
+    pub fn get(&self) -> u64 {
+        self.0.load(std::sync::atomic::Ordering::Relaxed)
+    }
+
+    #[inline(always)]
+    pub fn get_consistent(&self) -> u64 {
+        self.0.load(std::sync::atomic::Ordering::SeqCst)
+    }
+
+    #[inline(always)]
+    pub fn set(&self, value: u64) {
+        self.0.store(value, std::sync::atomic::Ordering::Relaxed)
+    }
+}
+
+pub type SharedConfigU32Value = SharedConfigValue<AtomicU32>;
+pub type SharedConfigU64Value = SharedConfigValue<AtomicU64>;
+
 /// Config that can be shared updated
 #[derive(Debug)]
 pub struct SharedReplicaConfig {
     pub base_dir: PathBuf,
-    pub index_max_bytes: SharedConfigValue,
-    pub index_max_interval_bytes: SharedConfigValue,
-    pub segment_max_bytes: SharedConfigValue,
-    pub flush_write_count: SharedConfigValue,
-    pub flush_idle_msec: SharedConfigValue,
-    pub max_batch_size: SharedConfigValue,
+    pub index_max_bytes: SharedConfigU32Value,
+    pub index_max_interval_bytes: SharedConfigU32Value,
+    pub segment_max_bytes: SharedConfigU32Value,
+    pub flush_write_count: SharedConfigU32Value,
+    pub flush_idle_msec: SharedConfigU32Value,
+    pub max_batch_size: SharedConfigU32Value,
     pub update_hw: bool, // if true, enable hw update
-    pub retention_seconds: SharedConfigValue,
+    pub retention_seconds: SharedConfigU32Value,
+    pub max_partition_size: SharedConfigU64Value,
 }
 
 impl From<ReplicaConfig> for SharedReplicaConfig {
     fn from(config: ReplicaConfig) -> Self {
         SharedReplicaConfig {
             base_dir: config.base_dir,
-            index_max_bytes: SharedConfigValue::new(config.index_max_bytes),
-            index_max_interval_bytes: SharedConfigValue::new(config.index_max_interval_bytes),
-            segment_max_bytes: SharedConfigValue::new(config.segment_max_bytes),
-            flush_write_count: SharedConfigValue::new(config.flush_write_count),
-            flush_idle_msec: SharedConfigValue::new(config.flush_idle_msec),
-            max_batch_size: SharedConfigValue::new(config.max_batch_size),
+            index_max_bytes: SharedConfigU32Value::new(config.index_max_bytes),
+            index_max_interval_bytes: SharedConfigU32Value::new(config.index_max_interval_bytes),
+            segment_max_bytes: SharedConfigU32Value::new(config.segment_max_bytes),
+            flush_write_count: SharedConfigU32Value::new(config.flush_write_count),
+            flush_idle_msec: SharedConfigU32Value::new(config.flush_idle_msec),
+            max_batch_size: SharedConfigU32Value::new(config.max_batch_size),
             update_hw: config.update_hw,
-            retention_seconds: SharedConfigValue::new(config.retention_seconds),
+            retention_seconds: SharedConfigU32Value::new(config.retention_seconds),
+            max_partition_size: SharedConfigU64Value::new(config.max_partition_size),
         }
     }
 }